Moved the render time tests accordingly
