+++ src/utils.rs
@@ -382,6 +382,27 @@ mod tests {
     use super::*;
 
     #[test]
+    fn test_500ms() {
+        assert_eq!(render_time(500_u128, true), "500ms")
+    }
+    #[test]
+    fn test_10s() {
+        assert_eq!(render_time(10_000_u128, true), "10s")
+    }
+    #[test]
+    fn test_90s() {
+        assert_eq!(render_time(90_000_u128, true), "1m30s")
+    }
+    #[test]
+    fn test_10110s() {
+        assert_eq!(render_time(10_110_000_u128, true), "2h48m30s")
+    }
+    #[test]
+    fn test_1d() {
+        assert_eq!(render_time(86_400_000_u128, true), "1d")
+    }
+
+    #[test]
     fn exec_mocked_command() {
         let result = exec_cmd("dummy_command", &[], Duration::from_millis(500));
         let expected = Some(CommandOutput {
